+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog Twitch::Bot
 
+## v4.1.1
+
+- [NEW] `CommandHandler` by default responds to `:user_message` events.
+
+## v4.1.0
+
+* [NEW] New `CommandHandler` class that simplifies building chat commands.
+
 ## v4.0.1
 
 * [FIXED] Fixed test crash due to incomplete DotEnv initialization.